+++ src/components/Form/Form.tsx
@@ -19,6 +19,7 @@ import { ClearFormButton } from "./ClearFormButton";
 import { FormLink } from "./Fields/FormLink/FormLink";
 import { TooltipItem } from "./Fields/TooltipItem/TooltipItem";
 import { MultiSelect } from "./Fields/MultiSelect/MultiSelect";
+import { UnorderedList } from "./Fields/UnorderedList/UnorderedList";
 
 interface IForm {
     onSubmit?: (data: any, e?: React.BaseSyntheticEvent<object, any, any>) => any;
@@ -60,5 +61,6 @@ export {
     RadioButton,
     FormLink,
     TooltipItem,
-    MultiSelect
+    MultiSelect,
+    UnorderedList
 }