+++ src/models/IFormInput.ts
@@ -55,6 +55,14 @@ export interface IMultiSelect extends IColumnPlacement {
     resetValue?: boolean;
 }
 
+export interface IUnorderedList extends IColumnPlacement {
+    textRows: string[];
+    name: string;
+    label: string;
+    inlineLabel: boolean;
+    className?: string;
+}
+
 export interface IFileUpload extends IFormInputBaseWithValidation, IFormInputValidation, IColumnPlacement {
     multiple?: boolean;
     allowedFileTypes?: string[];
@@ -91,6 +99,7 @@ export interface ICheckbox extends IFormInputBaseWithValidation, IColumnPlacemen
     id?: string;
     onChange?: (isChecked: boolean) => void;
     withColumn?: boolean;
+    tooltipDescription?: string;
 }
 
 export interface ICheckboxList extends IColumnPlacement {